feat(sim): add ExSIA pipeline and DIM64 RTL
Carry signed-64 results through ABI v3 and replace flattened DIM64 scheduling with a 16x16 tiled hierarchy.

Ultraworked with [omo](https://github.com/code-yeongyu/oh-my-openagent)

+++ sim/tests/common/runner.rs
@@ -4,7 +4,7 @@ use super::{golden_output, k_fragments, KBlockScaleMatrix, KFragment, Shape};
 
 pub struct RunResult {
     pub output: Vec<i32>,
-    pub expected: Vec<i32>,
+    pub expected: Vec<i64>,
     pub fragments: Vec<KFragment>,
     pub stats: Vec<TileStats>,
 }
